Add separate resources and test target for stress tests in CI/CD
